requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.3
bs4==0.0.2
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.7
djangorestframework==3.16.0
googlemaps==4.10.0
idna==3.10
python-dotenv==1.1.0
requests==2.32.3
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.13.0
urllib3==2.3.0
